feat(theme): configure use of own theme
We add or own theme through the sassdown options. This only concerns the css
file not yet the template file. To start we just increase the width of the gutter
to ensure that or widest breakpoint's behavior can be observed too. We also
make sure or css and js files are linked up correctly.