feat(styles): add global go-desktop breakpoint
We add global breakpoints to describe where layout changes from mobile
(as we style mobile first) to desktop views should happen. This is
mainly needed for the navigation configuration, when the off canvas
will be active its toggle will show up and when the navbar will be
displayed. This lets us control all this through one single variable.