feat(build): use new changelog generator
An improved changelog generator is implemented which uses a
new API and therefore breaks previous configurations. These
are rewritten to work with the new API. Also the bump, tag
and push logic is optimized to work better with this new
workflow.

BREAKING CHANGE: The new conventional-changelog API uses a renewed configuration
and will not work with previous versions. Therefore be careful using this
version's package.json with older configurations.