fix(js): accordionToggle lookup conditional
We need to look up the accordion toggle information conditionally since
an error is thrown if the element cannot be found. Therefore we first
check if an accordion toggle is on the page and only then extract the
information. For future stuff maybe think about adding all config stuff
to a global array then add this array to the body:before content.