feat(lists): add finalized list atom
The new mixin allows us to define all kinds of lists that are highly
configurable. From their selector to the icon used to the list position,
everything can easily be passed to the mixin and will be compiled to
a good form. Also we set the default list-style-position to inside
which allows us to style list indents better.