fix(scss): import scss normalize partial explicitly
Since node sass seems to be able to import scss and css partials now we need
to specify the scss partial in the normalize import explicitly for the sass
compile not to fail. Also we reposition the normalize styles so they appear
as the first rules in the final stylesheet.